+++ testsuite/src/aes.rs
@@ -1732,9 +1732,9 @@ mod tests {
             let start: u32 = DWT::cycle_count();
             let cipher = aes_gcm::Aes128Gcm::new(key.as_ref().into());
             let nonce = Nonce::from_slice(iv.as_ref().into());
-            let result_tag = cipher
+            let result_tag = unwrap!(cipher
                 .encrypt_in_place_detached(nonce, &gcm.aad, &mut buf[..gcm.pt.len()])
-                .unwrap();
+                .ok());
             total_elapsed += DWT::cycle_count().wrapping_sub(start);
 
             let result_tag: [u8; 16] = unwrap!(result_tag.try_into());