+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Added `Rtc::alarm_{a,b}` to get the alarm value.
+- Added `impl From<Alarm> for chrono::NaiveTime`.
+
 ## [0.5.0] - 2022-05-08
 ### Added
 - Added `set_sleep_clock` to GPIO ports to enable and disable clocks during sleep.